🎨 split up the adapter module
